Consolidate streaming parser tests into conformance tests
