+++ agent_sdks/cpp/CMakeLists.txt
@@ -102,11 +102,10 @@ target_link_libraries(a2ui_sdk PUBLIC nlohmann_json::nlohmann_json nlohmann_json
 add_executable(a2ui_tests
   tests/schema/test_catalog.cc
   tests/schema/test_schema_manager.cc
-  tests/parser/test_streaming_v08.cc
-  tests/parser/test_streaming_v09.cc
-  tests/parser/test_streaming_utils.cc
+  tests/parser/test_streaming_parser.cc
   tests/test_conformance.cc
 )
+
